refactor(metal): utils module for data format between metal (GPU) and arkworks (CPU)
